Prerequisite for bug 127
Since the changes in get.py in ops-restd are dependent on these tests and vice-versa, they need to be disabled and merged first. Only, then the changes in the ops-restd would be able checked-in and merged. After that I will make the changes in the tests to reanble them and upload them for review again.
As part of the bug fix empty optional columns will be removed from the "get" data. Since this is change on the REST implementation side, some tests will require modification to accomodate those changes.

Change-Id: Ib0b13f2b97b9f1497ce96e94b2f86e6c202f155a
Signed-off-by: Akshar Ranka <akshar.ranka@hpe.com>