add: dev: Schema for show-version-detail
We plan to display the git hash's for the GIT repos and
version numbers for other daemons.
Also we'll be displaying the source code location for each package.

Absorbed comments for removing the maxLength clause in the XML file
Removed the max limit for the number of entries in the table

Change-Id: I2e175d9b250c1f90e57c34f95b69e430e4e2aeb8
Signed-off-by: Sagar Bhanagay <sagar.bhanagay@hpe.com>