new: dev: Adding new table CPU_Rx to filter and copy multicast packets to CPU for further processing by respective daemons.
Updated as per the 4/12/2016 review comments.

Initial review, which was abandoned because of merge issues.
https://review.openswitch.net/#/c/4718/

This is initial table definition to filter and copy OSPF, LLDP, LCAP, etc
L2/L3 multicast packets to CPU.
Respective daemons will enter rows for the multicast/address type traffic
to be allowed. More columns and restriction can be added later depending on
further requirements.

Change-Id: Ic3802349a0f8aed7a7ca928c7f0b0e58bb9b9be2
Signed-off-by: aslammdkhan <aslam.khan@hpe.com>