chg: new: Schema changes for L2 MAC flush operations
The intent of putting this MAC flush table in the schema is to allow for flushing MAC addresses in bulk i.e. flushing
all MACs learnt on a VLAN or on a port.The vswitchd would read from this table and call the driver APIs/feature plugins
to achieve the MAC flush request. This can be a generic mechanism by which multiple features/daemons may request a bulk MAC flush.

Change-Id: I8a49bc89034f681633143195d40e00a193d5d6b7
Signed-off-by: shanmugamari <shanmugamari.s@hpe.com>