new: dev: Adding dynamic-category elements to route table
A new "valueCategory" element is added inside the "from" column
to identify when some columns in a row represent status or
configuration to let the REST interface change rows representing
static routes.

Columns in a table with "categoryFollows" refer to the column
containing the value to be used to identify how to handle it.

Change-Id: I64af81fd5b7167a891b4eae3f0f275de4c86804f
Signed-off-by: Francisco Rojas <francisco.rojas@hpe.com>
Signed-off-by: Javier Albornoz <javier.albornoz@hpe.com>