fix: dev: Taiga bug #680 NTP not working with REST
Root cause:
Because we have min and max defined for NTP_Association Table,
REST is unable to get information about the association which
has a zero or one weak reference to VRF table.

Solution:
Fixing the NTP_Association table to hold weak VRF reference but
have only ONE weak reference.

And this is as per the email on ops-dev 
http://lists.openswitch.net/pipermail/ops-dev/2016-March/002855.html

Change-Id: If60627a925dacac6d417e9c59b91d94c29ed47e0
Signed-off-by: srinsriv <srinivasan.srivatsan@hpe.com>