chg: dev: Adding new columns to summarize aggregation state
A new column is added to the interface table and the port table to
summarize the state of the aggregation.  This is done for two
to make it easier for users (protocols or humans) to identify the
LAG state without having to check individual member interfaces
or the hardware state.

Adding schema changes and feature documentation MD file.

TG-967

(cherry picked from commit 81becf26c6a1539c30d2cfaaa64f63a3bfefddac)

Change-Id: Ida2cebb7ae75e1bd5706ebb5f55c189c334a315f
Signed-off-by: Kattia Chaves <kchavesr@hpe.com>