chg: dev: Adding new columns to summarize aggregation state
A new column is added to the interface table and the port table to
summarize the state of the aggregation.  This is done for two
to make it easier for users (protocols or humans) to identify the
LAG state without having to check individual member interfaces
or the hardware state.

Change-Id: Ia4eb599294701d528dd88f5427edd22f19b2760b
Signed-off-by: Francisco Rojas <francisco.rojas@hpe.com>