chg: dev: update power_supplies schema
Added the following columns to power_supplies table: identity,
watts, and statistics.  Each of these columns are key value pairs
that contain multiple entries.  The identity column includes the
product name and serial number of the PSU.  The watts column
includes the total current and total maximum power values for the
PSU.  The statistics column includes the number of warnings and
number of failures experienced by the PSU.

Change-Id: I2e0ca1c34d9f65c7b950cf29a2abf7af8cf77f6e
Signed-off-by: Rachel Callison <rachel.callison@hpe.com>