chg: dev: Schema change to add column table-id in VRF table
The column table_id will hold the value of vrf id (say 0 to a maximum value) represented as an integer.
This is in addition to the usage of existing uuid, not as complementary. This ID will be dynamically generated by vrfmgrd.
Table_id is required for the following reasons:

1. Table_id is an integer in typical ASICs. We need to map from uuid to this ID every time when need to translate.
2. Unicast / multicast modules requires table_id to be mapped as integer for easy translation and mapping.
3. VRF_device investigation lead us to the need of remembering the table_id per vrf_device for route table programming. Table_id for default VRF will be 0.

The proposal is adding a table_id column. VRF aware applications can use this ID as a VRF index, or they can generate their own ID.

When a VRF (say vrf RED) gets deleted and added again, the same VRF-ID will not be allotted. The next possible number
in the free-list will be allotted. This will ensure that VRF-IDs will not be re-used immediately.

Change-Id: Ic9d7e0567da93b178ac8750532f653b2220671bc
Signed-off-by: Jayachandra Babu K <jayachandra.bab.k@hpe.com>