chg: dev: add kernel_state variable to port table
    Currently, there is no field in port table to indicate if the
port is really available in kernel, due to this, it is unclear of
provisioning. Provisioning failures are not being tracked as of now.
Lack of this causing many daemons need to implement retry logic.
With port pecking order, data path initialization is being finalized,
but control path is not considered. This changes are intending to
cover the gap and having kernel state maintained in port table.

Change-Id: I60609b8c562562b97574f124ccc6114a30aae81a
Signed-off-by: krishna <elluru.kri.mohan@hpe.com>