Add emptyValue for SNMP related columns in XML schema
The emptyValue key will be introduced in the new extended schema format,
to hold the assumed value for an optional column if not set. Currently,
this information is in plain text in some column's XML description. This
change adds the emptyValue as an XML tag attribute so that automation
tools to convert the schema to the new format pick it up when parsing the
XML schema.

Tags: chg, dev

Change-Id: I4a275dc3a81c9e84a3950968943b6805893fc02b
Signed-off-by: Eduardo Ramirez <eramireza@hpe.com>