Add emptyValue for ECMP related columns in XML schema
The emptyValue key will be introduced in the new extended schema format,
to hold the assumed value for an optional column if not set. Currently,
this information is in plain text in some column's XML description. This
change adds the emptyValue as an XML tag attribute so that automation
tools to convert the schema to the new format pick it up when parsing the
XML schema.

Tags: chg, dev

Change-Id: Ic5f83ab562e99af42d85c91fef529711b3861ef2
Signed-off-by: Eduardo Ramirez <eramireza@hpe.com>