Add emptyValue for Authentication related columns in XML schema
The emptyValue key will be introduced in the new extended schema format,
to hold the assumed value for an optional column if not set. Currently,
this information is in plain text in some column's XML description. This
change adds the emptyValue as an XML tag attribute so that automation
tools to convert the schema to the new format pick it up when parsing
the
XML schema.

Tags: chg, dev

Change-Id: I88b1f4351ead6c20f4c10c6e47b85ebc401a58f0
Signed-off-by: Eduardo Ramirez <eramireza@hpe.com>