Add emptyValue for columns from Nexthop table in XML schema
The emptyValue key will be introduced in the new extended schema format,
to hold the assumed value for an optional column if not set. Currently,
this information is in plain text in some column's XML description. This
change adds the emptyValue as an XML tag attribute so that automation
tools to convert the schema to the new format pick it up when parsing
the
XML schema.

Tags: chg, dev

Change-Id: I16260754517e364a27027d4340d5778e4e12e9d8
Signed-off-by: Eduardo Ramirez <eramireza@hpe.com>