Add emptyValue for columns from Bridge table in XML schema
The emptyValue key will be introduced in the new extended schema format,
to hold the assumed value for an optional column if not set. Currently,
this information is in plain text in some column's XML description. This
change adds the emptyValue as an XML tag attribute so that automation
tools to convert the schema to the new format pick it up when parsing
the
XML schema.

Tags: chg, dev

Change-Id: I3c18e5da2d1b24cf924162068f5e3e4d1e9e18dc
Signed-off-by: Eduardo Ramirez <eramireza@hpe.com>