Add missing constraints for Mirror table in XML schema
The documentation in the XML schema for the Mirror table references an
'inactive' operation state that is not included in its enum.
The emptyValue key will be introduced in the new extended schema format,
to hold the assumed value for an optional column if not set. Currently,
this information is in plain text in some column's XML description. This
change adds the emptyValue as an XML tag attribute so that automation
tools to convert the schema to the new format pick it up when parsing
the
XML schema.

Tags: chg, dev

Change-Id: Ia2e1ab2a751e1772e5447ad6991a532cc9fe91c1
Signed-off-by: Eduardo Ramirez <eramireza@hpe.com>