Modification of schema
1. MAC table
The old vlan column has a constraint of being in range of 1 to 4094.
Creating a new row for MAC table without setting vlan column and setting
new column mac_vlan, would result in constraint violation, as value of
vlan column be by default would tried to be set as 0.

Tags: fix, dev
TG-1175

Change-Id: I3853b5f7b47803b287747354fecf079e10d3bce2
Signed-off-by: Harshvardhan <harshvardhan@hpe.com>